@@ -8,61 +8,59 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "prepare": "husky install",
+    "prepare": "husky",
     "lint:format": "npx --no-install lint-staged && npx --no-install pretty-quick --staged"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.4",
-    "@angular/cdk": "^15.1.4",
-    "@angular/common": "^15.1.4",
-    "@angular/compiler": "^15.1.4",
-    "@angular/core": "^15.1.4",
-    "@angular/forms": "^15.1.4",
-    "@angular/material": "^15.1.4",
-    "@angular/platform-browser": "^15.1.4",
-    "@angular/platform-browser-dynamic": "^15.1.4",
-    "@angular/router": "^15.1.4",
-    "bootstrap": "^4.6.0",
+    "@angular/animations": "^17.2.4",
+    "@angular/cdk": "^17.2.2",
+    "@angular/common": "^17.2.4",
+    "@angular/compiler": "^17.2.4",
+    "@angular/core": "^17.2.4",
+    "@angular/forms": "^17.2.4",
+    "@angular/material": "^17.2.2",
+    "@angular/platform-browser": "^17.2.4",
+    "@angular/platform-browser-dynamic": "^17.2.4",
+    "@angular/router": "^17.2.4",
+    "bootstrap": "^5.3.3",
     "rxjs": "^7.8.0",
     "tslib": "^2.5.0",
-    "zone.js": "^0.12.0"
+    "zone.js": "^0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.5",
-    "@angular-eslint/builder": "^15.2.1",
-    "@angular-eslint/eslint-plugin": "^15.2.1",
-    "@angular-eslint/eslint-plugin-template": "^15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "^15.2.1",
-    "@angular/cli": "^15.1.5",
-    "@angular/compiler-cli": "^15.1.4",
-    "@commitlint/cli": "^12.1.4",
-    "@commitlint/config-conventional": "^12.1.4",
-    "@types/jasmine": "^3.6.11",
-    "@types/node": "^14.17.0",
-    "@typescript-eslint/eslint-plugin": "^5.43.0",
-    "@typescript-eslint/parser": "^5.43.0",
-    "eslint": "^8.28.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-import": "^2.27.5",
+    "@angular-devkit/build-angular": "^17.2.3",
+    "@angular-eslint/builder": "^17.2.1",
+    "@angular-eslint/eslint-plugin": "^17.2.1",
+    "@angular-eslint/eslint-plugin-template": "^17.2.1",
+    "@angular-eslint/schematics": "17.2.1",
+    "@angular-eslint/template-parser": "^17.2.1",
+    "@angular/cli": "^17.2.3",
+    "@angular/compiler-cli": "^17.2.4",
+    "@commitlint/cli": "^19.0.3",
+    "@commitlint/config-conventional": "^19.0.3",
+    "@types/jasmine": "^5.1.4",
+    "@types/node": "^18.19.22",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "eslint": "^8.53.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jasmine": "^4.1.3",
-    "eslint-plugin-jsdoc": "^40.0.0",
+    "eslint-plugin-jsdoc": "^48.2.1",
     "eslint-plugin-prefer-arrow": "latest",
-    "eslint-plugin-rxjs": "^5.0.2",
-    "husky": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "karma": "~6.3.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-spec-reporter": "0.0.32",
-    "lint-staged": "^10.5.4",
-    "prettier": "^2.8.4",
-    "pretty-quick": "^3.1.3",
-    "protractor": "~7.0.0",
-    "ts-node": "^10.9.1",
-    "typescript": "~4.9.5"
+    "eslint-plugin-rxjs": "^5.0.3",
+    "husky": "^9.0.11",
+    "jasmine-core": "^5.1.2",
+    "karma": "^6.4.3",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage": "^2.2.1",
+    "karma-jasmine": "^5.1.0",
+    "karma-spec-reporter": "^0.0.36",
+    "lint-staged": "^15.2.2",
+    "prettier": "^3.2.5",
+    "pretty-quick": "^4.0.0",
+    "typescript": "~5.3.3"
   },
   "lint-staged": {
     "*.{ts,html}": [
