@@ -1,11 +1,11 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 
-import { EMPTY, Observable, of, throwError } from 'rxjs';
-import { map } from 'rxjs/operators';
+import {EMPTY, Observable, of} from 'rxjs';
+import {map} from 'rxjs/operators';
 
-import { Product } from './product.interface';
+import {Product} from './product.interface';
 
-import { ApiService } from '../core/api.service';
+import {ApiService} from '../core/api.service';
 
 @Injectable({
   providedIn: 'root',
