@@ -8,7 +8,7 @@ describe('CartComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [CartComponent],
+      imports: [CartComponent],
     }).compileComponents();
   });
 
